requirements.txt
aio_pika
dynaconf
fastapi
pydantic
python-dotenv
python-multipart
redis
slowapi
uvicorn